Move CA1309 to its new place - no functional change.
