+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/Microsoft.ApiDesignGuidelines.CSharp.Analyzers.props
@@ -11,6 +11,8 @@
         $(CodeAnalysisRuleSetOverrides);
         -Microsoft.Design#CA1001;
         -Microsoft.Design#CA1019;
+
+        -Microsoft.Globalization#CA1309;
       </CodeAnalysisRuleSetOverrides>
    </PropertyGroup>
 </Project>
